Documents/final year project /project/ml-models/requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
statsmodels==0.14.0
mysql-connector-python==8.1.0
mlxtend==0.22.0
